refactor: moved lulo actions to hook
